requirements.txt
openai
groq
google-genai
python-dotenv
fastapi
tqdm
uvicorn
pydantic
requests
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
langchain-core==0.3.68
langchain-google-genai==2.1.6
langsmith==0.4.4
datasets
minio
python-multipart
psycopg2-binary
pandas
sqlalchemy